+++ clover/common/plugin/postman.py
@@ -1,0 +1,6 @@
+
+from clover.common.plugin import Pipeline
+
+class PostmanPlugin(): pass
+
+class Postman(Pipeline): pass
